+++ intersect_test.go
@@ -209,3 +209,27 @@ func TestUnion(t *testing.T) {
 	is.Equal(result4, []int{0, 1, 2})
 	is.Equal(result5, []int{})
 }
+
+func TestWithout(t *testing.T) {
+	is := assert.New(t)
+	result1 := Without([]int{0, 2, 10}, 0, 1, 2, 3, 4, 5)
+	result2 := Without([]int{0, 7}, 0, 1, 2, 3, 4, 5)
+	result3 := Without([]int{}, 0, 1, 2, 3, 4, 5)
+	result4 := Without([]int{0, 1, 2}, 0, 1, 2)
+	result5 := Without([]int{})
+	is.Equal(result1, []int{10})
+	is.Equal(result2, []int{7})
+	is.Equal(result3, []int{})
+	is.Equal(result4, []int{})
+	is.Equal(result5, []int{})
+}
+
+func TestWithoutEmpty(t *testing.T) {
+	is := assert.New(t)
+	result1 := WithoutEmpty([]int{0, 1, 2})
+	result2 := WithoutEmpty([]int{1, 2})
+	result3 := WithoutEmpty([]int{})
+	is.Equal(result1, []int{1, 2})
+	is.Equal(result2, []int{1, 2})
+	is.Equal(result3, []int{})
+}